moved project creation into the project model
