+++ front-end/src/lib/models/project.ts
@@ -1,4 +1,5 @@
 import axios from "@/lib/axios";
+import { AxiosError } from "axios";
 
 export interface Project {
   id: number;
@@ -8,4 +9,23 @@ export interface Project {
 export async function getProjects() {
   const projectsResponse = await axios.get<Project[]>("/projects");
   return projectsResponse.data;
+}
+
+export async function createProject(name: string) {
+  try {
+    await axios.post("/projects", {
+      name,
+    });
+  } catch (error) {
+    if (!(error instanceof AxiosError) || error.response?.status !== 409) {
+      if (error instanceof AxiosError) {
+        console.error(error.message);
+        console.error(error.response?.data);
+      }
+
+      return { error: "Er is iets misgegaan" };
+    }
+
+    return { error: "Unexpected error" };
+  }
 }
